Try reverting race-condition fixes
Why:

The fixes are not addressing all race-conditions (namely the
race-condition between trial fetch and sampling) and makes the fetching
of trials less efficient. Therefore, if the problem of duplicates was
mainly due to using to small uniform distribution for sampling seeds, it
is better to revert these fixes and use the more efficient version.